+++ unifyt/__init__.py
@@ -23,7 +23,7 @@ from unifyt.serialization import (
     quantity_decoder,
 )
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"
 __all__ = [
     "Quantity",
     "Unit",